Each kind of node would implement toAst method
