+++ lib/io/ast_writer.dart
@@ -2,8 +2,6 @@ import 'web.dart' if (dart.library.io) 'desktop.dart';
 
 import 'foundation.dart';
 import '../models/document.dart';
-import '../models/node.dart';
-import '../protos/ast.pb.dart';
 
 class AstWriter {
   AstWriter(this._doc, this._handle);
@@ -18,15 +16,5 @@ class AstWriter {
 extension DocumentAstExportable on Document {
   List<int> toAst() {
     return root.toAst();
-  }
-}
-
-extension RootNodeAstExportable on RootNode {
-  List<int> toAst() {
-    final result = Root();
-    // TODO: build the full Root with calls/streams and sub values.
-    final call = Call();
-    result.calls.add(call);
-    return result.writeToBuffer();
   }
 }
